fix: Encodeのテストで任意の値を扱えるように変更
