+++ iosTraining-ReeenTests/iosTraining_ReeenTests.swift
@@ -68,10 +68,14 @@ final class iosTraining_ReeenTests: XCTestCase {
     }
 
     func testEncodeRequestParameters() {
-        let weatherService = WeatherService()
+        let encoder = WeatherEncoder()
+        let date = DateComponents(calendar: Calendar(identifier: .gregorian), timeZone: TimeZone(secondsFromGMT: 0), year: 2023, month: 7, day: 7, hour: 12, minute: 1, second: 1).date!
+        let dummyRequestParameters = RequestParameters(area: "tokyo", date: date)
+        let expectData = #"{"area":"tokyo","date":"2023-07-07T12:01:01Z"}"#
+
         do {
-            let encodedRequest = try weatherService.encodeRequestParameters()
-            XCTAssertNotNil(encodedRequest)
+            let encodedRequest = try encoder.encodeRequestParameters(dummyRequestParameters)
+            XCTAssertEqual(encodedRequest, expectData)
         } catch {
             XCTFail("Encoding failed with error: \(error)")
         }
@@ -79,7 +83,7 @@ final class iosTraining_ReeenTests: XCTestCase {
 
 
     func testDecodeWeatherInfo() {
-        let weatherService = WeatherService()
+        let decoder = WeatherDecoder()
         let dummyWeatherInfo = """
             {
                 "weather_condition": "sunny",
@@ -88,7 +92,7 @@ final class iosTraining_ReeenTests: XCTestCase {
             }
             """
         do {
-            let weatherData = try weatherService.decodeWeatherInfo(dummyWeatherInfo)
+            let weatherData = try decoder.decodeWeatherInfo(dummyWeatherInfo)
             XCTAssertEqual(weatherData.weatherCondition, "sunny")
             XCTAssertEqual(weatherData.maxTemperature, 30)
             XCTAssertEqual(weatherData.minTemperature, 20)
